Define default `docker` superuser
With 6.0.0 we default passwords can not be used in prod mode.  We use
the file realm to define a `docker` user with advanced privileges to
allow operations on the http API.

Additionally, x-pack 6.0.0 requires transport.ssl, so we rely on the
`certificate` verification mode to ignore validity of IP/hostname on
the pre-generated certs. As typically Docker networks are
encrypted (e.g. the overlay network in Swarm) and predicting IP
addresses is hard in Docker, this is something we can use for a smooth
getting started experience.