Change logic for detecting staging artifacts
Looking for a `-` in the complete version_tag is error prone as there
can be artifacts like `6.0.0-alpha1` not related to a BC candidate.

Expose both `ELASTIC_VERSION` and `STAGING_BUILD_NUM` in the jinja2
template used for building the Dockerfile and also the
`docker-compose.yml` for the tests though the Makefile.